requirements.txt
nltk==3.4.3
PyYAML==5.1.1
